requirements.txt
discord.py==2.3.2
fastapi==0.110.0
uvicorn==0.29.0
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.0
jinja2==3.1.3
python-multipart==0.0.9